chore(net): remove OP stack bootnodes
Remove optimism.rs and flatten bootnodes/ directory module into a
single bootnodes.rs since only Ethereum bootnodes remain.

Amp-Thread-ID: https://ampcode.com/threads/T-019c42ce-7362-770c-8b2d-2306e72e1260
